requirements.txt
api4jenkins
PyYAML
paramiko
scp